Fix compilation errors
This fixes the compilation issue of wrong reference for the
configuration variable.

also, in order to avoid a package-wide singleton for the configuration,
this instead reads the configuration into a variable which is then
propagated into the given sub-commands. This makes the sub-command
functions easier to test in the future.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>