Add golangci-lint configuration and fix warnings
This fixes a whole lot of warnings and removes dead code from the
codebase. The intent is to make this easier to digest.

In the end, there is still a bunch of dead code... like the pagination
code, for instance. However, I left that in place since it might come in
handy in the near future.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>