Remove resource update, simplify schema
To simplify things further for MVP, resource update has been removed
and the schema has been substantially simplified. Built-in roles are
good, and we should support them, but just to keep things simple for
now they have been removed. We can reintroduce them once we're
confident in the design so far.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>