Add API authentication to permissions-api
In general, it is expected that services require authentication even
when called directly instead of through a gateway. This commit
replaces the prior authentication strategy of setting the subject URN
as the bearer token value with JWT bearer token authentication. To
make local development easier, this commit also adds a
mock-oauth2-server service to the dev container environment and an
example config that uses it for JWT validation.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>