Authenticate API requests (#53)
* Add API authentication to permissions-api

In general, it is expected that services require authentication even
when called directly instead of through a gateway. This commit
replaces the prior authentication strategy of setting the subject URN
as the bearer token value with JWT bearer token authentication. To
make local development easier, this commit also adds a
mock-oauth2-server service to the dev container environment and an
example config that uses it for JWT validation.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Explicitly parse URNs rather than use strings package

The urnx package provides a convenient way for us to parse
Infratographer URNs across services. This commit updates
permissions-api to use it for subject and resource URNs instead of
string manipulation logic.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove mock-oauth2-server-health check

This check isn't being used anywhere and breaks in CI. This commit
removes it.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Use logger in place of fmt.Println for URN errors

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Run `go mod tidy` and make toolbox a direct dependency

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>