Add README with initial concepts and persist them throughout the repo
This adds an initial README that describes the concepts that Permissions
API will cover. This allows us to have more consistent conversations in
the future about functionality and logic within the project.

This also persists the concepts and replaces the relevant wording in
functions and variables throughout the repository.

Note that the worker has not yet been touched, as we need to revisit how
that piece will work.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>