Invert subject and resource relationship for relationships API
SpiceDB's (and Zanzibar's) notion of resource and subject are the
inverse of what one would typically think of, especially when looking
at schema definitions. This commit updates the relationships API to
allow for the definition of relationships for multiple subjects on a
single resource, which matches schemas and resources as defined more
closely.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>