Add Helm chart and Dockerfile (#66)
* Rename config and command items to use permissions-api consistently

The code as written has different variations on permissions-api, such
as permissionapi, permission-api, permission api, etc. This commit
updates the code and dev container configuration to consistently refer
to the service as permissions-api.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add Helm chart for permissions-api

In order to deploy permissions-api in Kubernetes effectively, we need
a Helm chart. This commit adds an official permissions-api Helm chart
to the repository leveraging the Bitnami common library chart.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add healthcheck endpoint

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add Dockerfile

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add CI steps for image builds and Helm chart releases

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add some useful utilities to dev container Docker image

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>