Use Trivy more effectively (#68)
* Use Trivy more effectively

The current image scanning workflow leads to Trivy scans being run
twice because the test action is run on both PRs and pushes to
main. To rectify this, a new security action has been added in this
commit that only runs on PRs to scan both the permissions-api Git
repository and Docker image. Additionally, repo scanning has been
added to the image-build action and a typo in the image tag to be
scanned has been fixed.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Set exit code for Trivy jobs since they seem to run now

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix Trivy findings in dev container Dockerfile

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Disable config scanner until it gets more better

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Adjust whitespace in dev container Dockerfile

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>