Hardcode role and tenant actions into schema
Roles and tenants can also be acted upon like any other
resource. However, they're not necessarily something we want to be
configured, as we need some minimum set of resources as well as
actions that can (or can't) be taken on resources. This commit adds
relations and permissions for roles and tenants to resolve this.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>