Better actions (#77)
* Separate actions and tenant actions in resource type

Some actions can be taken on a resource type directly, where others
can only be taken on their tenant context (e.g., get vs list). This
commit updates the ResourceType definition to separate actions and
tenant actions, and updates the loadbalancer resource accordingly.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Hardcode role and tenant actions into schema

Roles and tenants can also be acted upon like any other
resource. However, they're not necessarily something we want to be
configured, as we need some minimum set of resources as well as
actions that can (or can't) be taken on resources. This commit adds
relations and permissions for roles and tenants to resolve this.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add loadbalancer and role resources to query engine

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix broken logic that resulted in malformed roles when listing

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix broken test

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix linter issue

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>