Add NATS consumer for resource lifecycle events
All resource lifecycle events are expected to flow to permissions-api
through NATS. This commit introduces a NATS JetStream consumer based
on a prior implementation of NATS worker logic in permissions-api that
listens for events on resource types. Currently it only prints event
types and does not do anything based on the type provided for each
resource.

Additionally, this commit adds NATS-specific development tooling and
updates the Makefile to perform some of the tool fetching/building
logic that was previously in the dev container Dockerfile to reduce
build times in CI.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>