Add NATS consumer for resource lifecycle events (#81)
* Add NATS consumer for resource lifecycle events

All resource lifecycle events are expected to flow to permissions-api
through NATS. This commit introduces a NATS JetStream consumer based
on a prior implementation of NATS worker logic in permissions-api that
listens for events on resource types. Currently it only prints event
types and does not do anything based on the type provided for each
resource.

Additionally, this commit adds NATS-specific development tooling and
updates the Makefile to perform some of the tool fetching/building
logic that was previously in the dev container Dockerfile to reduce
build times in CI.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add support for lifecycle events

This commit adds support for handling the create, update, and delete
lifecycle events in permissions-api. Currently update and delete have
a stub implementation, while create will create actual relationships
in SpiceDB.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Don't ignore nats directory

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add support for update and delete lifecycle events

This commit adds support for update and delete lifecycle events on
resources. Update is a non-atomic operation that is equivalent to a
delete event followed by create event due to limitations in SpiceDB.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix error handling and logging for unknown event types

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add tests, fix miscellaneous logical bugs

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Use Ack instead of Term when handling errors

Calling Term on a message instructs the server to stop redelivering
the message but does not mark it as delivered. This is not really what
we want - ultimately, we still received a message and should
acknowledge its receipt instead of dropping it.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Run "go mod tidy"

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix linting issues

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add comment clarifying subscriptions for resource types

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Polish ack/nak semantics a bit

Explicitly acking/naking messages everywhere in the subscriber feels
like a bad pattern. This commit updates the NATS subscriber to nak a
message if an error has occurred and ack it otherwise.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Listen on SIGINT and SIGTERM

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Rework client logic to have defaults for everything

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>