correct relationship check direction
Previously we were checking if the destination resource had a relation
rather than the targeted resource having a relation to the destination.

This resulted in the following not working:

```yaml
resourcetypes:
  - name: folder
    relationships:
      - relation: parent
        targettypenames:
          - folder
  - name: file
    relationships:
      - relation: parent
        targettypenames:
          - folder
```

When creating a relationship previously, relating a folder to a parent
folder would work since they happen to have the same resource type.
However relating a file to a parent folder would fail due to us checking
if folder has a parent relationship type name for file.

Signed-off-by: Mike Mason <mimason@equinix.com>