role delete endpoint should be just role id
spicedb doesn't have a great way to find all objects which relate to a
given id and so the resource id which a role was assigned to was required.

This however is not a great user experience, so instead we discover the
resource and its assigned relations and then process the deletion with
the discovered information.

Signed-off-by: Mike Mason <mimason@equinix.com>