Remove relationship create/delete handlers (#174)
* Remove relationship create/delete handlers

NATS is the interface being used for relationship management
now. relationshipCreate and relationshipDelete are redundant and also
hard to make secure, since this part of the API is more of a
management of the SpiceDB graph itself instead of management of
resources that are on the SpiceDB graph. This commit removes both of
them and their respective API routes.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove unused types and functions

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>