Update mermaid template (#175)
* Update mermaid template with a few fixes

The mermaid template in the schema command wasn't showing relationship
actions for resources and showed one-to-many relationships between
resources. Technically, we don't keep track of constraints on number
of relationships, so to reflect that in the ER diagram this commit
updates the edges between nodes to reflect that. It also adds related
actions to union types.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Rename "perm" to "action" in mermaid template

This commit renames "perm" to "action" in the mermaid template to make
the intent of the attribute a little clearer to consumers.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>