use full consistent reads
This switches us to fully consistent reads of relationships which
ensures after any action has been taken a subsequent read request will
return the most up to date information.

This resolves an issue seen when directly after creating a role, a
client may fail to assign a subject to the newly created role.

Signed-off-by: Mike Mason <mimason@equinix.com>