wrap errors for easier handling in external services
Services wanting to check if the error returned was an auth error or
some other error could get a number of auth related errors.

This wraps the permissions and auth errors into easily comparable root
errors for easier checks.

Additionally this corrects the incorrectly returned echo errors which
were missing for PermissionDenied errors which resulted in 500 errors.

Signed-off-by: Mike Mason <mimason@equinix.com>