Add design proposal and implementation for ZedTokens table
While we have some guarantees that we try to make around consistency
in permissions-api, particularly around ZedTokens, these were not
explicitly spelled out anywhere. This commit adds a design proposal
for using CRDB itself to store ZedTokens for resources, rather than
relying on streaming from SpiceDB or lookups in NATS KV.

The motivations for the CRDB-based approach are that it grants us the
consistency guarantees we want while also allowing for more
flexibility than NATS provides around the location of a given leader
for coordinating reads, while also removing a dependency from the
permissions-api critical path.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>