Simplify role fetching logic in query engine (#282)
* Simplify role fetching logic in query engine

Prior implementations of the query engine fetched role information
such as the owning resource ID directly from SpiceDB, as it was the
only data store available. With the introduction of CRDB, that is no
longer the case and the CRDB SQL table should be considered the
authoritative source of most role data. This commit updates the query
engine to fetch role resource owner ID and other data from the SQL DB
whenever possible, getting rid of some obscure failure modes that
occur when a role has no associated actions.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix error type in RBAC v2 tests

As described.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Wrap LockRoleForUpdate in a method to return non-DB errors

As described.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Fix incorrect error in role update test case

As described.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>