properly handle echo http wrapped errors
Errors may be wrapped by a specific echo http error by a handler.
In some cases (such as canceled context) this might report as a 500
error, however this error might've been triggered by the requestor.

To simplify handling this scenario, we will unwrap the http error,
preserving the original message only if the error is a context canceled
error and the request context has been canceled.

Signed-off-by: Mike Mason <mimason@equinix.com>